added a readme and moved files to the package level
