+++ code401challenges/src/main/java/code401challenges/linkedList/README.MD
@@ -1,0 +1,4 @@
+# Linked List
+### Author: James Dansie
+
+This lab implements a singly linked list. This list includes the methods of contains, insert head, insert tail, and toString. This lab also has jUnit tests for each of the methods. https://www.geeksforgeeks.org/implementing-a-linked-list-in-java-using-class/ is a very helpful guide. This list has been rewritten to not use the static method though. So far only accepts ints into the list.
