+++ code401challenges/src/test/java/code401challenges/linkedList/LinkedListTest.java
@@ -1,4 +1,4 @@
-package code401challenges;
+package code401challenges.linkedList;
 
 import org.junit.Test;
 
@@ -7,7 +7,26 @@ import static org.junit.Assert.*;
 public class LinkedListTest {
 
     @Test
-    public void testListCreationEmpty() {
+    public void testCreate(){
+        LinkedList list = new LinkedList();
+        assertNull(list.head);
+    }
+
+    @Test
+    public void testHead() {
+        LinkedList list = new LinkedList();
+        list.insertTail(2);
+        list.insertTail(3);
+        list.insertHead(1);
+
+        assertEquals(
+                "Head should be 1",
+                1,
+                list.head.value);
+    }
+
+    @Test
+    public void testListCreationInsert() {
         LinkedList list = new LinkedList();
         list.insertTail(2);
         list.insertTail(3);
@@ -16,8 +35,7 @@ public class LinkedListTest {
         assertEquals(
                 "This should contain 1, 2, 3",
                 "Linked List: 1 2 3 ",
-                list.toString())
-        ;
+                list.toString());
     }
 
     @Test
@@ -30,8 +48,7 @@ public class LinkedListTest {
         assertEquals(
                 "This should contain 1, 2, 3",
                 true,
-                list.contains(2))
-        ;
+                list.contains(2));
     }
 
     @Test
@@ -44,7 +61,6 @@ public class LinkedListTest {
         assertEquals(
                 "This should contain 1, 2, 3",
                 false,
-                list.contains(10))
-        ;
+                list.contains(10));
     }
 }
